// lib/game-config/generated.ts
export const cards = [
  {
    "name": "R1",
    "rarity": "R",
    "image": "/images/cards/player/R1/card.webp"
  },
  {
    "name": "R2",
    "rarity": "R",
    "image": "/images/cards/player/R2/card.webp"
  },
  {
    "name": "R3",
    "rarity": "R",
    "image": "/images/cards/player/R3/card.webp"
  },
  {
    "name": "R4",
    "rarity": "R",
    "image": "/images/cards/player/R4/card.webp"
  },
  {
    "name": "SR1",
    "rarity": "SR",
    "image": "/images/cards/player/SR1/card.webp"
  },
  {
    "name": "SR2",
    "rarity": "SR",
    "image": "/images/cards/player/SR2/card.webp"
  },
  {
    "name": "SR3",
    "rarity": "SR",
    "image": "/images/cards/player/SR3/card.webp"
  },
  {
    "name": "SR4",
    "rarity": "SR",
    "image": "/images/cards/player/SR4/card.webp"
  },
  {
    "name": "SSR1",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR1/card.webp"
  },
  {
    "name": "SSR2",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR2/card.webp"
  },
  {
    "name": "SSR3",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR3/card.webp"
  },
  {
    "name": "SSR4",
    "rarity": "SSR",
    "image": "/images/cards/player/SSR4/card.webp"
  },
  {
    "name": "UR1",
    "rarity": "UR",
    "image": "/images/cards/player/UR1/card.webp"
  },
  {
    "name": "UR2",
    "rarity": "UR",
    "image": "/images/cards/player/UR2/card.webp"
  },
  {
    "name": "UR3",
    "rarity": "UR",
    "image": "/images/cards/player/UR3/card.webp"
  }
] as const;
// ...
export const battleEnemies = {
  "1": {
    "id": 1,
    "name": "Enemy 1",
    "image": "/images/cards/enemy/enemy1/card.webp",
    "attackCounter": 5
  },
  "2": {
    "id": 2,
    "name": "Enemy 2",
    "image": "/images/cards/enemy/enemy2/card.webp",
    "attackCounter": 4
  },
  "3": {
    "id": 3,
    "name": "Enemy 3",
    "image": "/images/cards/enemy/enemy3/card.webp",
    "attackCounter": 7
  },
  "4": {
    "id": 4,
    "name": "Enemy 4",
    "image": "/images/cards/enemy/enemy4/card.webp",
    "attackCounter": 5
  },
  "5": {
    "id": 5,
    "name": "Enemy 5",
    "image": "/images/cards/enemy/enemy5/card.webp",
    "attackCounter": 8
  },
  "6": {
    "id": 6,
    "name": "Enemy 6",
    "image": "/images/cards/enemy/enemy6/card.webp",
    "attackCounter": 6
  },
  "7": {
    "id": 7,
    "name": "Enemy 7",
    "image": "/images/cards/enemy/enemy7/card.webp",
    "attackCounter": 4
  },
  "8": {
    "id": 8,
    "name": "Enemy 8",
    "image": "/images/cards/enemy/enemy8/card.webp",
    "attackCounter": 9
  },
  "9": {
    "id": 9,
    "name": "Enemy 9",
    "image": "/images/cards/enemy/enemy9/card.webp",
    "attackCounter": 9
  },
  "10": {
    "id": 10,
    "name": "Enemy 10",
    "image": "/images/cards/enemy/enemy10/card.webp",
    "attackCounter": 15
  },
  "11": {
    "id": 11,
    "name": "Enemy 11",
    "image": "/images/cards/enemy/enemy11/card.webp",
    "attackCounter": 5
  },
  "12": {
    "id": 12,
    "name": "Enemy 12",
    "image": "/images/cards/enemy/enemy12/card.webp",
    "attackCounter": 8
  },
  "13": {
    "id": 13,
    "name": "Enemy 13",
    "image": "/images/cards/enemy/enemy13/card.webp",
    "attackCounter": 11
  }
} as const;

export const enemySelectionGroups = [
  {
    "phase": "normal",
    "rounds": "2,3,4,6,7,8,9",
    "group": "common",
    "weight": 60.0,
    "enemyIds": [
      1,
      2,
      4,
      7,
      11
    ]
  },
  {
    "phase": "normal",
    "rounds": "2,3,4,6,7,8,9",
    "group": "rare",
    "weight": 27.0,
    "enemyIds": [
      3,
      5,
      6,
      8,
      9
    ]
  },
  {
    "phase": "normal",
    "rounds": "2,3,4,6,7,8,9",
    "group": "special",
    "weight": 3.0,
    "enemyIds": [
      10
    ]
  },
  {
    "phase": "chance",
    "rounds": "1,5,10",
    "group": "common",
    "weight": 35.0,
    "enemyIds": [
      1,
      2,
      4,
      7,
      11
    ]
  },
  {
    "phase": "chance",
    "rounds": "1,5,10",
    "group": "rare",
    "weight": 60.0,
    "enemyIds": [
      3,
      5,
      6,
      8,
      9
    ]
  },
  {
    "phase": "chance",
    "rounds": "1,5,10",
    "group": "special",
    "weight": 5.0,
    "enemyIds": [
      10
    ]
  },
  {
    "phase": "extra",
    "rounds": "11+",
    "group": "reset",
    "weight": 55.0,
    "enemyIds": [
      12
    ]
  },
  {
    "phase": "extra",
    "rounds": "11+",
    "group": "boss",
    "weight": 20.0,
    "enemyIds": [
      13
    ]
  },
  {
    "phase": "extra",
    "rounds": "11+",
    "group": "mixed",
    "weight": 25.0,
    "enemyIds": [
      1,
      3,
      10
    ]
  }
] as const;